+++ resources/[dg]/dg-imager/typescript/src/server/constant.ts
@@ -1,0 +1,2 @@
+export const IMGUR_ID = '5ed58555618e037';
+export const MINIO_BUCKET_ID = 'dg-image-storage';